@@ -33,13 +33,15 @@
     "chai": "4.3.6",
     "js-base64": "^3.7.2",
     "jwt-encode": "^1.0.1",
+		"mocha": "^10.1.0",
     "mocha-remote-cli": "^1.6.1",
     "mocha-remote-server": "^1.6.0",
     "puppeteer": "^14.3.0",
-    "@realm/app-importer": "^0.1.0",
+    "@realm/app-importer": "*",
     "realm-web": "*",
-    "webpack": "^5.73.0",
-    "webpack-dev-server": "^4.9.2"
+    "webpack": "^5.88.0",
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^5.9.0"
   },
   "peerDependencies": {
     "mocha-remote-client": "^1.6.1"
@@ -47,16 +49,14 @@
   "devDependencies": {
     "@types/chai": "^4.3.1",
     "@types/fs-extra": "^9.0.13",
-    "@types/mocha": "^8",
-    "@types/node-fetch": "^2.6.2",
+    "@types/mocha": "^10.0.1",
+    "@types/node-fetch": "^2.6.4",
     "@types/puppeteer": "^5.4.6",
-    "@types/webpack": "^5.28.0",
-    "@types/webpack-env": "^1.17.0",
+    "@types/webpack-env": "^1.18.1",
     "fs-extra": "^10.1.0",
-    "html-webpack-plugin": "^5.5.0",
-    "mongodb-realm-cli": "^1.3.2",
+    "html-webpack-plugin": "^5.5.3",
     "node-fetch": "^3.2.5",
-    "ts-loader": "^9.3.0",
-    "webpack-cli": "^4.9.2"
+    "ts-loader": "^9.4.3",
+    "webpack-cli": "^5.1.4"
   }
 }
